snapzy/services/ai/captioning-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
redis==5.0.4
httpx==0.27.0
pillow==10.3.0